chore(vats): separate sim, client, chain boot behaviors
  - update boot-psm.js now that manifest.js is gone
  - move BootstrapManifest type to lib-boot.js

chore(vats): reorg boot-chain, chain-behaviors

 - restore test-boot.js tests

 - static types for test-boot.js and boot-*.js
   - prune argvByRole
 - no more export *
 - lib-boot: early check for missing behavior

 - basic-behaviors.js:
   - organize imports a bit
   - all behaviors return Promise<void>
     use await rather than return
   - export BASIC_BOOTSTRAP_PERMITS
 - boot-chain.js:
   - destructure non-behaviors out of import *
   - hoist modules
 - chain-behaviors: factor BASIC_BOOTSTRAP_PERMITS out
   of SHARED_CHAIN_BOOTSTRAP_MANIFEST
