+++ packages/vats/decentral-core-config.json
@@ -3,7 +3,7 @@
   "defaultReapInterval": 1000,
   "vats": {
     "bootstrap": {
-      "sourceSpec": "@agoric/vats/src/core/boot.js",
+      "sourceSpec": "@agoric/vats/src/core/boot-chain.js",
       "creationOptions": {
         "critical": true
       }
