+++ packages/vats/index.js
@@ -2,7 +2,6 @@
 // Ambient types
 import './src/core/types.js';
 
-export * from './src/core/boot.js';
 // eslint-disable-next-line import/export -- no named exports
 export * from './src/types.js';
 export * from './src/nameHub.js';
