--- packages/vats/src/core/behaviors.js
@@ -1,4 +1,0 @@
-// @ts-check
-export * from './basic-behaviors.js';
-export * from './chain-behaviors.js';
-// We exclude sim-behaviors.js from this list because it should not be used in production.
