+++ packages/vats/src/core/boot-sim.js
@@ -1,69 +1,46 @@
 // @ts-check
-import * as simBehaviors from '@agoric/inter-protocol/src/proposals/sim-behaviors.js';
-import {
-  CLIENT_BOOTSTRAP_MANIFEST,
-  CHAIN_BOOTSTRAP_MANIFEST,
-  SIM_CHAIN_BOOTSTRAP_MANIFEST,
-} from './manifest.js';
+import { SIM_CHAIN_BOOTSTRAP_PERMITS } from '@agoric/inter-protocol/src/proposals/sim-behaviors.js';
+import * as simBehaviorsPlus from '@agoric/inter-protocol/src/proposals/sim-behaviors.js';
+import * as basicBehaviorsPlus from './basic-behaviors.js';
+import { SHARED_CHAIN_BOOTSTRAP_MANIFEST } from './chain-behaviors.js';
+import * as chainBehaviorsPlus from './chain-behaviors.js';
 
-import * as behaviors from './behaviors.js';
-import * as clientBehaviors from './client-behaviors.js';
-import * as utils from './utils.js';
 import { makeBootstrap } from './lib-boot.js';
 
-const { Fail } = assert;
+export const MANIFEST = {
+  ...SHARED_CHAIN_BOOTSTRAP_MANIFEST,
+  ...SIM_CHAIN_BOOTSTRAP_PERMITS,
+};
 
-// Choose a manifest based on runtime configured argv.ROLE.
-const roleToManifest = harden({
-  chain: CHAIN_BOOTSTRAP_MANIFEST,
-  'sim-chain': SIM_CHAIN_BOOTSTRAP_MANIFEST,
-  client: CLIENT_BOOTSTRAP_MANIFEST,
-});
-const roleToBehaviors = harden({
-  'sim-chain': { ...behaviors, ...simBehaviors },
-  // copy to avoid trying to harden a module namespace
-  client: { ...clientBehaviors },
-});
+const {
+  BASIC_BOOTSTRAP_PERMITS: _b,
+  PowerFlags: _p,
+  makeMyAddressNameAdminKit: _m,
+  ...basicBehaviors
+} = basicBehaviorsPlus;
+const {
+  CHAIN_BOOTSTRAP_MANIFEST: _c,
+  SHARED_CHAIN_BOOTSTRAP_MANIFEST: _s,
+  ...chainBehaviors
+} = chainBehaviorsPlus;
+const { SIM_CHAIN_BOOTSTRAP_PERMITS: _sc, ...simBehaviors } = simBehaviorsPlus;
+const behaviors = { ...basicBehaviors, ...chainBehaviors, ...simBehaviors };
 
 /**
- * Build root object of the bootstrap vat.
+ * Build root object of the bootstrap vat for the simulated chain.
  *
  * @param {{
  *   D: DProxy,
  *   logger: (msg) => void,
  * }} vatPowers
  * @param {{
- *   argv: { ROLE: string },
- *   bootstrapManifest?: Record<string, Record<string, unknown>>,
  *   coreProposalCode?: string,
  * }} vatParameters
  */
 export const buildRootObject = (vatPowers, vatParameters) => {
-  const {
-    argv: { ROLE },
-    bootstrapManifest,
-  } = vatParameters;
-  console.debug(`${ROLE} bootstrap starting`);
+  console.debug(`sim bootstrap starting`);
 
-  const bootManifest = bootstrapManifest || roleToManifest[ROLE];
-  const bootBehaviors = roleToBehaviors[ROLE] || behaviors;
-  bootManifest || Fail`no configured bootstrapManifest for role ${ROLE}`;
-  bootBehaviors || Fail`no configured bootstrapBehaviors for role ${ROLE}`;
-
-  const modules = {
-    clientBehaviors: { ...clientBehaviors },
-    simBehaviors: { ...simBehaviors },
-    behaviors: { ...behaviors },
-    utils: { ...utils },
-  };
-
-  return makeBootstrap(
-    vatPowers,
-    vatParameters,
-    bootManifest,
-    behaviors,
-    modules,
-  );
+  return makeBootstrap(vatPowers, vatParameters, MANIFEST, behaviors, {});
 };
 
 harden({ buildRootObject });
