+++ packages/vats/src/core/boot-solo.js
@@ -1,28 +1,38 @@
 // @ts-check
-import * as simBehaviors from '@agoric/inter-protocol/src/proposals/sim-behaviors.js';
-import {
-  CLIENT_BOOTSTRAP_MANIFEST,
-  CHAIN_BOOTSTRAP_MANIFEST,
-  SIM_CHAIN_BOOTSTRAP_MANIFEST,
-} from './manifest.js';
-
-import * as behaviors from './behaviors.js';
-import * as clientBehaviors from './client-behaviors.js';
-import * as utils from './utils.js';
 import { makeBootstrap } from './lib-boot.js';
 
-const { Fail } = assert;
+import * as basicBehaviorsPlus from './basic-behaviors.js';
+import * as chainBehaviorsPlus from './chain-behaviors.js';
+import { CLIENT_BOOTSTRAP_MANIFEST } from './client-behaviors.js';
+import * as clientBehaviorsPlus from './client-behaviors.js';
+import * as utils from './utils.js';
 
-// Choose a manifest based on runtime configured argv.ROLE.
-const roleToManifest = harden({
-  chain: CHAIN_BOOTSTRAP_MANIFEST,
-  'sim-chain': SIM_CHAIN_BOOTSTRAP_MANIFEST,
-  client: CLIENT_BOOTSTRAP_MANIFEST,
+export const MANIFEST = CLIENT_BOOTSTRAP_MANIFEST;
+
+// XXX export basicBehaviors record from basic-behaviors.js?
+const {
+  BASIC_BOOTSTRAP_PERMITS: _b,
+  PowerFlags: _p,
+  makeMyAddressNameAdminKit: _m,
+  ...basicBehaviors
+} = basicBehaviorsPlus;
+const {
+  CHAIN_BOOTSTRAP_MANIFEST: _c,
+  SHARED_CHAIN_BOOTSTRAP_MANIFEST: _s,
+  ...chainBehaviors
+} = chainBehaviorsPlus;
+const { CLIENT_BOOTSTRAP_MANIFEST: _cb, ...clientBehaviors } =
+  clientBehaviorsPlus;
+const behaviors = harden({
+  ...basicBehaviors,
+  ...chainBehaviors,
+  ...clientBehaviors,
 });
-const roleToBehaviors = harden({
-  'sim-chain': { ...behaviors, ...simBehaviors },
-  // copy to avoid trying to harden a module namespace
-  client: { ...clientBehaviors },
+
+const modules = harden({
+  clientBehaviors: { ...clientBehaviors },
+  behaviors: { ...behaviors },
+  utils: { ...utils },
 });
 
 /**
@@ -39,31 +49,9 @@ const roleToBehaviors = harden({
  * }} vatParameters
  */
 export const buildRootObject = (vatPowers, vatParameters) => {
-  const {
-    argv: { ROLE },
-    bootstrapManifest,
-  } = vatParameters;
-  console.debug(`${ROLE} bootstrap starting`);
+  console.debug(`solo client bootstrap starting`);
 
-  const bootManifest = bootstrapManifest || roleToManifest[ROLE];
-  const bootBehaviors = roleToBehaviors[ROLE] || behaviors;
-  bootManifest || Fail`no configured bootstrapManifest for role ${ROLE}`;
-  bootBehaviors || Fail`no configured bootstrapBehaviors for role ${ROLE}`;
-
-  const modules = {
-    clientBehaviors: { ...clientBehaviors },
-    simBehaviors: { ...simBehaviors },
-    behaviors: { ...behaviors },
-    utils: { ...utils },
-  };
-
-  return makeBootstrap(
-    vatPowers,
-    vatParameters,
-    bootManifest,
-    behaviors,
-    modules,
-  );
+  return makeBootstrap(vatPowers, vatParameters, MANIFEST, behaviors, modules);
 };
 
 harden({ buildRootObject });
